Align directory name to namespace
