+++ src/tim/transform/ops/conv2d_layout_inference.h
@@ -27,8 +27,8 @@
 #include "tim/vx/ops/conv2d.h"
 
 #include "src/tim/vx/operation_private.h"
-#include "src/tim/layout_infer/permute_vector.h"
-#include "src/tim/layout_infer/ops/op_layout_inference.h"
+#include "src/tim/transform/permute_vector.h"
+#include "src/tim/transform/ops/op_layout_inference.h"
 
 namespace tim {
 namespace transform {
