+++ src/tim/transform/ops/elementwise_layout_inference.h
@@ -26,8 +26,8 @@
 
 #include "tim/vx/ops/elementwise.h"
 
-#include "src/tim/layout_infer/ops/op_layout_inference.h"
-#include "src/tim/layout_infer/permute_vector.h"
+#include "src/tim/transform/ops/op_layout_inference.h"
+#include "src/tim/transform/permute_vector.h"
 #include "src/tim/vx/operation_private.h"
 
 namespace tim {
