+++ src/tim/transform/ops/op_layout_inference.cc
@@ -23,7 +23,7 @@
  *****************************************************************************/
 
 #include "op_layout_inference.h"
-#include "src/tim/layout_infer/permute_vector.h"
+#include "src/tim/transform/permute_vector.h"
 #include "src/tim/vx/operation_private.h"
 #include "tim/vx/ops/transpose.h"
 #include "src/tim/vx/type_utils.h"
